--- src/main/kotlin/io/gijung/aichatbot/auth/JwtGenerateDTO.kt
@@ -1,6 +1,0 @@
-package io.gijung.aichatbot.auth
-
-data class JwtGenerateDTO(
-    val userId: String,
-    val userRole: String,
-)
