+++ src/main/kotlin/io/gijung/aichatbot/auth/infrastructure/jwt/ClockHolder.kt
@@ -1,0 +1,5 @@
+package io.gijung.aichatbot.auth.infrastructure.jwt
+
+fun interface ClockHolder {
+    fun now(): Long
+}
