+++ src/main/kotlin/io/gijung/aichatbot/auth/infrastructure/jwt/SystemClockHolder.kt
@@ -1,0 +1,8 @@
+package io.gijung.aichatbot.auth.infrastructure.jwt
+
+import org.springframework.stereotype.Component
+
+@Component
+class SystemClockHolder : ClockHolder {
+    override fun now(): Long = System.currentTimeMillis()
+}
